customisable cars and wheels
I made it so you are able to customise your car body and wheels and then changed the race logic to follow the wheels and body. The check mark in garage scene should only show up if the body and wheels match what's chosen. I removed the old cars as they were only one sprite. In garage scene, alter the wheel dimension in the car definitions so the wheels fit the car body.