Created the perfect start system and altered notification styles
Added a new PerfectStart system to RaceScene and refactored Car and NitrousMiniGame to display notifications with backgrounds and outlines, and ensures notifications are properly reset on race restart.

These notifications pop up when a player does: a perfect shift, activates/fails the nitrous minigame or has a slow/perfect start.